requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
requests==2.32.3
six==1.16.0
sympy==1.13.1
torch==2.5.1
typing_extensions==4.12.2
urllib3==2.2.3
